agents/search/requirements.txt
langgraph>=0.2.0
python-dotenv>=1.0.1
tavily-python>=0.4.0
